Update patch set 3
Patch Set 3:

Awesome!
lgtm

A nit though is that the CL title is a bit weird so I think that should be changed. It describes a problem rather than what is done in the CL.
What about:
Remove the ability for EchoRemover to override the render vs capture alignment

Patch-set: 3
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>