Update patch set 2
Patch Set 2:

(1 comment)

Niels, the next CL I have in line uses this. It needs to translate the last-received and last-decoded together. If it gets both, it can use them. If either is missing from the map, it ignores the LossNotification message. So I don't think this is premature optimization.

Patch-set: 2
Reviewer: Gerrit User 5147 <5147@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0