Add RTP sequence number to TransportFeedbackObserver::AddPacket()
With this change, both the normal RTP and the transport-wide sequence
numbers are propagated with with AddPacket() call.

The intent with this is that RtpTransportControllerSend can hold a
mapping from transport-wide to rtp sequence numbers, and then let the
RTP modules know via a callback when packets have been received by the
remote end.

Bug: webrtc:8975
Change-Id: I6a24fc6282cbb041393752d39593c2867b242192