Implement RTCRemoteInboundRtpStreamStats.
<This CL will be split up into smaller CLs>

This CL implements the "remote-inbound-rtp" stats object for both
audio and video. These are metrics based on Report Blocks, which gives
feedback from the remote receiving side about the "health" of the RTP
stream corresponding to a local sending stream (which has a
corresponding "outbound-rtp" stats object).

This CL implements the following metrics:
- timestamp
- ssrc
- kind
- transportId
- codecId
- packetsLost
- jitter
- localId
- roundTripTime

These are most, but not all, of the RTCRemoteInboundRtpStreamStats
dictionary minus "burst", "gap" and some "packet" metrics.
https://w3c.github.io/webrtc-stats/#remoteinboundrtpstats-dict*

Notably this CL does *not* add:
- packetsReceived: We need to subtract the sequence number from the
  initial sequence number and handle integer wrap-around.
- fractionLost: Requires additional calculations.

Bug: webrtc:10455
Change-Id: I54432a81ac0cacf0c2f55d2a33aa6c006263bc1e