Update patch set 21
Patch Set 21: Commit-Queue+1

(44 comments)

Thanks for the review!

Pushing back on one point: I want to keep the name <DataCodec> around, and use it as the
template argument for the DataContentDescription class, for 2 reasons:

1) when I tried to define 2 classes that were both based on MediaContentDescriptionImpl<DataCodec>, I got an obscure template error. That's what made me rename the codec. Typedef'ing it seems to work fine.

2) even then, this exists so that Hangouts can have their code unmodified. If their code uses
DataCodec, it makes sense to use that name in the API, even when it's a typedef.

Otherwise, I hope it all works still.

Patch-set: 21
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+1