Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > Hi, I don't object to adding back configurability of the retransmissions, if that's useful for applications. But I'm not so fond of exposing the RetransmissionMode bit masks, and I'd prefer construction time settings over setter methods.
> > > 
> > > Which values of the uint8_t settings do you see a use for? Do you need to modify settings mid-stream?
> > 
> > The only one we need is retransmit all.
> 
> Oh, sorry, and no: we don't need to dynamically change it, setting it in the constructor (directly or through config) would be totally fine.

Thanks for the explanation. Then I'd suggest making a separate cl adding a bool or enum to the RtpConfig class, which is passed to the RtpVideoSender constructor, and translate it to what the RTPSenderVideo (apologies for this naming) needs there.

Patch-set: 4
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>