Update patch set 2
Patch Set 2: Commit-Queue+1

> Patch Set 1: Code-Review+1
> 
> There seems to be one SetRateAllocation method left, in api/video_codecs/test/video_encoder_software_fallback_wrapper_unittest.cc. Delete that one too?
> 
> According to git grep 'SetRates.*,' '*.h', there also seems to be a two-argument SetRates method left in a few places, including modules/video_coding/codecs/test/videoprocessor.h, modules/video_coding/utility/frame_dropper.h, modules/video_coding/utility/frame_dropper.h, and an OnSetRates in video/frame_encode_timer.h. Probably best left for a followup cl, but would be nice to clean up all of those.

All of those are just internal helper methods, not part of this interface.
I can clean some of them up in a follow-up cl if you like, but prefer to keep this on as short as possible.

Patch-set: 2
Reviewer: Gerrit User 5524 <5524@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+1