Update patch set 3
Patch Set 3: Code-Review+1

(4 comments)

Looks good overall, but I'm not quite sure I understand the configuration requirements.  It seems to me that the scrolling config asks the caller for a lot of redundant information, so I'd consider whether there's a way to infer some of that information from elsewhere.  But if that's not possible, this seems ok.

Patch-set: 3
Reviewer: Gerrit User 5599 <5599@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1