Update patch set 1
Patch Set 1:

(1 comment)

Change looks good to me.

But it seems a bit brittle to me to have responsibility for the |network_report_| member pingpong between threads. Would it be possible to instead pass a reference to an RTCStatsReport between threads with PostTask, and either delete the member or access it strictly from one of the threads?

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>