Update patch set 3
Patch Set 3:

(3 comments)

This cl is getting a bit large, and is not so easy to review. When you are confident you have the big picture right, it would help to split it into pieces, e.g., taking care on one MediaChannel method and its callers at a time. Hopefully that should give smaller and focused cls that are easy to review.

Patch-set: 3
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>