Update patch set 5
Patch Set 5:

> Hmm, so does webrtc wait until if the current bitstream refers the previous frame in decoding order? It sounds inefficient.

No, I meant, that if some frame is missing, even if webrtc receives the next one, it won't be sent for decoding (because it's missing a reference). This is what frame_buffer2 does. It waits for the next decodable frame.

> Could you elaborate? One receiver will get different substs of SVC stream at the same time?
> Does this combine drop spartial index?

E.g. one receiver has enough bandwidth to receive high bandwidth fullHD stream. The second receiver has bad connection and receives only spatial layers 0,1. It gets trickier with K-SVC, there the first receiver gets only SL2 for P-frames and SL0,SL1,SL2 for I-frames, while the second receiver gets SL1 for P-frames and SL0, SL1 for I-frames.

The extra layers are filtered out by the backend according to the available downlink capacity. This is why we can't write the index on the send side, or it will have to be rewritten by the backend, so again, no need to do it.

Patch-set: 5
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>