Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (2 comments)
> > 
> > > Patch Set 3:
> > > 
> > > Adding Danil.
> > > 
> > > In general I'd like to avoid structs with modifiable public members, but I see that the dependency struct is not being introduced in this change. I'd prefer a constructor that sets the state once and avoid the possibility of having multiple states possible throughout the instances lifetime.
> > > 
> > > Danil - wdyt about having a constructor for the dependency class and having the members be const? (think |std::string const foo_| and not |const std::string foo_|).
> > 
> > That topic seems out of scope of this change.
> > [on that topic I think making members const would kill the idea of this struct. This struct is not suppose to live long or stored as a member. It is there just to create MediaEngine. Same pattern is used in CreateModularPeerConnectionFactory and CreatePeerConnection. ]
> 
> The reason I think it's relevant is because this CL adds a dependency on that design.
> Is there something in place that guarantees that instances of the struct aren't long lived, used on more than one thread or have values changed throughout their lifetime?

I do not understand what misuse you are suggesting to avoid,
what is wrong with this way of using builder pattern.

this is a struct, according to style guide we are using, structs are passive, they provide no guarantees. (https://google.github.io/styleguide/cppguide.html#Structs_vs._Classes)

Without imagination I do not see how this struct can be useful for anything other than for intended usage: Create, fill, move to CreateMediaEngine function, forget.

But if someone can imagine scenario where it might be helpful to keep this struct longer and/or change it's values from different threads, then it would be up to the user, up to the active class(es) to provide guarantees it is done in a thread-safe way.

Patch-set: 3
Reviewer: Gerrit User 5019 <5019@58829da1-049c-39fb-b951-ebdcd0984767>