Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > (2 comments)
> > > 
> > > > Patch Set 3:
> > > > 
> > > > Adding Danil.
> > > > 
> > > > In general I'd like to avoid structs with modifiable public members, but I see that the dependency struct is not being introduced in this change. I'd prefer a constructor that sets the state once and avoid the possibility of having multiple states possible throughout the instances lifetime.
> > > > 
> > > > Danil - wdyt about having a constructor for the dependency class and having the members be const? (think |std::string const foo_| and not |const std::string foo_|).
> > > 
> > > That topic seems out of scope of this change.
> > > [on that topic I think making members const would kill the idea of this struct. This struct is not suppose to live long or stored as a member. It is there just to create MediaEngine. Same pattern is used in CreateModularPeerConnectionFactory and CreatePeerConnection. ]
> > 
> > The reason I think it's relevant is because this CL adds a dependency on that design.
> > Is there something in place that guarantees that instances of the struct aren't long lived, used on more than one thread or have values changed throughout their lifetime?
> 
> I do not understand what misuse you are suggesting to avoid,
> what is wrong with this way of using builder pattern.
> 
> this is a struct, according to style guide we are using, structs are passive, they provide no guarantees. (https://google.github.io/styleguide/cppguide.html#Structs_vs._Classes)
> 
> Without imagination I do not see how this struct can be useful for anything other than for intended usage: Create, fill, move to CreateMediaEngine function, forget.
> 
> But if someone can imagine scenario where it might be helpful to keep this struct longer and/or change it's values from different threads, then it would be up to the user, up to the active class(es) to provide guarantees it is done in a thread-safe way.

I may be overly paranoid with this, but I've refactored quite a bit of code in WebRTC and a common thing that becomes a problem, are public member variables. Whatever the initial thinking may have been when a struct was introduced, if it's easy to not be aware of the design and/or to break it, it can happen. Especially for types that are a part of an API and not internal to WebRTC.

In this case, an implementation of MediaEngineInterface, could easily have a member variable of type MediaEngineDependencies. That in itself would make the instance long lived.

My question has to do with the design constraints/assumptions and what, if anything has been done to make sure that developers are aware of them.

A good thing with a constructor is that you know when it's called, what parameters it needs and under what conditions it runs. The builder pattern is about convenience for the developer. There's a bit of a conflict between that pattern and RAII, another well known principle, because you construct an intermediary object that potentially has multiple states - unless you design the intermediary struct also following RAII rules, but that would re-introduce the inconvenience the builder pattern is meant to avoid in the first place :)

In any case, I'm not saying that there's any misuse going on here or that something wrong is being done. There's no comment for the deps class that clarifies how it's supposed to be used or not supposed to use, so I'm just asking about the design rather than assuming.

Patch-set: 3
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>