Update patch set 4
Patch Set 4:

Yes, that is part of what I was trying to imply.

I'll take another look at this tomorrow and if there is a reason to take a look at the constructor itself, we can do so separately.

Patch-set: 4
Tag: mailMessageId=<CANDN_yUiVp2v0CKocmxJK+JyCvcNQMxJSvnhOMtZFbEsuAtHcw@mail.gmail.com>