Update patch set 3
Patch Set 3:

> Look at WebRtcKeyValueConfig, I think this does the thing you want, it currently lives in api/transport/ but probably should be moved to something like api/field_trial or so as it's not really transport specific.

Thanks for the comment. I did take a look there. I think FieldTrialManager is a better name. It is more clear that it is an evolution of the field trial system rather than an entirely separate configuration system.

I am happy to replace usages of WebRtcKeyValueConfig in this or another CL.  

What are your thoughts?

Patch-set: 3
Reviewer: Gerrit User 8773 <8773@58829da1-049c-39fb-b951-ebdcd0984767>