Enforce DetectNumCpus returns 1 as default value.
It is possible for sysconf to return -1 in case of error [1], even
if this is unlikely to happen the code is more readable if we move
the default enforcement to the return line. This way the reader
doesn't have to read through the preprocessor #ifs because it will
be clear that if number_of_cores is 0 or less than 0, we will default
to 1.

Bug: None
Change-Id: I294684e7380b12cda55ec8d6c7a90e132dc3af85