Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)

I will revert this part of the change to keep the same interface as neteq. It does not make sense to increase the reference counter in neteq_quality_test to then send the reference into neteq without doing anything with it. Neteq will keep a copy of the pointer and increase the reference counter.

Patch-set: 5
Reviewer: Gerrit User 7967 <7967@58829da1-049c-39fb-b951-ebdcd0984767>