Update patch set 3
Patch Set 3:

(4 comments)

One worry is that it looks like old versions of libwebrtc will just parrot back whatever header extension IDs are in the offer. So if a newer client offers id > 14, it won't see the header extension removed in the answer and try sending I think? Maybe there's a condition lower in the code that will avoid sending two byte header extensions if a=extmap is not present in the answer?

Patch-set: 3
Reviewer: Gerrit User 5039 <5039@58829da1-049c-39fb-b951-ebdcd0984767>