In FrameEncodeMetadataWriter don't clear known bitrate on Reset.
Reset() is called each time the encoder is reconfigured, but then it
happens the target bitrate isn't reset in encoder. So it might produce a
frame before next bitrate estimate is propagated to the metadata writer.
The incorrect zero bitrate would be treated as a paused encoder and would
cause metadata to be dropped.

Also, added unittest for that scenario at VideoStreamEncoder level.

Bug:webrtc:10460
Change-Id: I28024a527f1fb8474b172e2c5c2394fd38d69a07