Update patch set 1
Patch Set 1:

(1 comment)

Have you considered instead adding a Builder constructor with a const VideoFrame& as argument? That would solve the problem, and let you write, e.g.,

  rotated_frame = VideoFrame::Builder(frame)
          .set_video_frame_buffer(...)
          .set_rotation(kVideoRotation_0).build();

Not that familiar with the builder pattern, though, so not sure if it's a good idea.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>