Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)
> 
> oprypin@ - The last CL in this refactoring will remove the jni_package variable: https://chromium-review.googlesource.com/c/chromium/src/+/1674254. I'm not familiar with the webrtc project structure - am I correct to assume that submitting that CL in chromium is safe without verifying downstream/subrepo webrtc CLs have been submitted?

It is "safe" in that it can't immediately break us, but if it's a "bad" change then we are unable to roll in new revisions of Chromium until that's fixed.

You can go to a Chromium CL and "CHOOSE TRYJOBS" and run WebRTC's tryjobs there to find such breakages. Sometimes these are unavoidable though, and we have to do something about them on WebRTC's side.

Patch-set: 6