Update patch set 2
Patch Set 2:

If nothing has been decoded, NetEq will only generate zero samples as PLC, right? That is something the decoder PLC can do as well, and it should know very well if anything has been decoded or not. That is, I think the onus of deciding if something has been decoded or not, and maybe also to provide the zero samples, should be on the decoder.

Let me know if you think I'm wrong.

Patch-set: 2
Reviewer: Gerrit User 5100 <5100@58829da1-049c-39fb-b951-ebdcd0984767>