Stop using unnecessary gclient vars
My understanding is that this purely cargo-culting. We never actually use the variables, nor do we need the "Three lines of non-changing comments", because we roll everything at once anyway.

I also remove the 'chromium_git' variable because the indirection is not helpful and makes the file harder to follow.

I went even further and concatenated all strings. Again, perhaps in the past this was used to more easily update the file with automated tools, but that is also not necessary now.

Verified: `gclient revinfo` output remains exactly the same.

Change-Id: Ib8fc33b0f11f53fab0fed9b9b578db964fb70bdc
Bug: None