Replace the implementation of `GetContributingSources()` on the audio side.
This change replaces the `ContributingSources`-implementation of `GetContributingSources()` and `GetSynchronizationSources()` on the audio side with the spec-compliant `SourceTracker`-implementation.

The most noticeable impact is that the per-frame dictionaries are now updated when frames are delivered to the RTCRtpReceiver's MediaStreamTrack rather than when RTP packets are received on the network.

This change is almost identical to the previous video side change at:

  https://webrtc-review.googlesource.com/c/src/+/143177

Bug: webrtc:10545
Change-Id: Ife7f08ee8ca1346099b7466837a3756947085fc5