Update patch set 9
Patch Set 9:

Did you look into what it would take to change this to be built lazily, and only actually write to buffer_ at the point where we need it (most likely when we put the packet on the wire)? Current design is a bit strange.

Patch-set: 9
Reviewer: Gerrit User 5525 <5525@58829da1-049c-39fb-b951-ebdcd0984767>