RtpPacketSender for sending packets, RtpPacketPacer for pacer control
The PacedSender is being reworked and will need an interface so we can
inject different implementatins of it.
Previously, we intended the RtpPacketPacer interface in the RTP module
to do this. That does not seem to work well instead this CL:

* Moves the old RtpPacketSender to where RtpPacketPacer currently
  resides, and includes both InsertPacket and EnqueuePacket.
  This interface is intended to only handle the task of sending packets,
  and so AccountForAudioPackets is moved out of it.
* Introduces a new RtpPacketPacer interface inside the pacing module.
  This interface handles the details of _how_ the packets should be
  paced, such as pacing rates/account for audio/max queue length etc

Implementation details, such as how to register processing threads,
will be delegated solely to the implementation.

Some minor cleanups are included here.

Change-Id: I150b1a6262306d99e3f9d5f0b4afdb16a50e5ad8