Make new pacer code path default-on
This make the path, where packets are directly owned by PacedSender
rather that being temporarily put in the RtpPacketHistory, the default.

Functionally, this should essentially be a noop, with only minor timing
differences.

The old code-path will stay around for a short while and then be
removed once we are certain there are no regressions.

Bug: webrtc:10633
Change-Id: If1114a7bd66bf212a1ed611d0eb4ffcac235a67a