Add keys that identify certain time series in event log visualizer.
For stream-related data, the key will typically be the SSRC. In general,
the meaning of the key depends on the chart. For example, we might want
to use the transaction ID to follow ICE pings.

Also add y-axis labels for catagorical data, i.e. the mapping between
numerical enum values and string labels.

Bug:webrtc:10623
Change-Id: Ic9c3da9a3874f479c07412f394a774ae90fd3d7e