Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> How long has the new code path been enabled? What makes us think it's safe to turn it on now?

It is not enabled, this CL enables it.

Note that https://webrtc-review.googlesource.com/c/src/+/146218 needs to be landed first.

It's been running in tests, and I've done a bunch of manual testing.
I don't think this change really merits an A/B test, that's why this CL just changes the default so that we can kill it with a flag if we discover issues in a later stage.

Patch-set: 1
Reviewer: Gerrit User 5524 <5524@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0